+++ docs/changelog.md
@@ -10,6 +10,7 @@
 - Use dataclasses to describe Source
 - Add param `max_rows` to `create_table` (import only part of a table, all
   plugins are supported)
+- `import_from_source` now supports compressed files (and so all CLI commands)
 
 ### Plugins
 ### Command-Line Interface